Moved the class to a separate file
