+++ netmiko/huawei/__init__.py
@@ -1,4 +1,5 @@
-from netmiko.huawei.huawei import HuaweiSSH, HuaweiVrpv8SSH, HuaweiOLTSSH
-from netmiko.huawei.huawei import HuaweiTelnet, HuaweiOLTTelnet
+from netmiko.huawei.huawei import HuaweiSSH, HuaweiVrpv8SSH
+from netmiko.huawei.huawei import HuaweiTelnet
+from netmiko.huawei.huawei_olt import HuaweiOLTSSH, HuaweiOLTTelnet
 
 __all__ = ["HuaweiSSH", "HuaweiVrpv8SSH", "HuaweiTelnet", "HuaweiOLTTelnet", "HuaweiOLTSSH"]
