+++ tests/etc/responses.yml.example
@@ -259,17 +259,17 @@ sophos_sfos:
   multiple_line_output: ""
 
 huawei_olt_telnet:
-  base_prompt: MA5608T_10.210.0.12
-  router_prompt: MA5608T_10.210.0.12>
-  enable_prompt: MA5608T_10.210.0.12#
+  base_prompt: MA5801_1.2.3.4
+  router_prompt: MA5801_1.2.3.4>
+  enable_prompt: MA5801_1.2.3.4#
   version_banner: "Uptime is"
   multiple_line_output: "Uptime is"
   interface_ip: "Uptime is"
 
 huawei_olt:
-  base_prompt: MA5608T_10.210.0.12
-  router_prompt: MA5608T_10.210.0.12>
-  enable_prompt: MA5608T_10.210.0.12#
+  base_prompt: MA5801_1.2.3.4
+  router_prompt: MA5801_1.2.3.4>
+  enable_prompt: MA5801_1.2.3.4#
   version_banner: "Uptime is"
   multiple_line_output: "Uptime is"
   interface_ip: "Uptime is"
